+++ providers/google/src/main/scala/com/guizmaii/distances/providers/google/utils/Implicits.scala
@@ -1,18 +1,20 @@
 package com.guizmaii.distances.providers.google.utils
 
 import cats.MonadError
-import cats.effect.Async
+import cats.effect.Concurrent
 import com.google.maps.PendingResult
 
 private[google] object Implicits {
 
   implicit final class PendingResultOps[T](val request: PendingResult[T]) {
-    def asEffect[F[_]: Async]: F[T] =
-      Async[F].async { cb =>
+    def asEffect[F[_]: Concurrent]: F[T] =
+      Concurrent[F].cancelable { cb =>
         request.setCallback(new PendingResult.Callback[T] {
           override def onResult(result: T): Unit     = cb(Right(result))
           override def onFailure(e: Throwable): Unit = cb(Left(e))
         })
+
+        Concurrent[F].delay(request.cancel())
       }
   }